Rename `SwiftScan` directory into `DependencyScan`
