+++ lib/CMakeLists.txt
@@ -19,6 +19,7 @@ add_subdirectory(ASTSectionImporter)
 add_subdirectory(Basic)
 add_subdirectory(ClangImporter)
 add_subdirectory(Demangling)
+add_subdirectory(DependencyScan)
 add_subdirectory(Driver)
 add_subdirectory(Frontend)
 add_subdirectory(FrontendTool)
@@ -36,7 +37,6 @@ add_subdirectory(PrintAsObjC)
 add_subdirectory(RemoteAST)
 add_subdirectory(Sema)
 add_subdirectory(Serialization)
-add_subdirectory(SwiftScan)
 add_subdirectory(SwiftDemangle)
 add_subdirectory(SwiftReflection)
 add_subdirectory(SwiftRemoteMirror)
