+++ lib/DependencyScan/DependencyScanningTool.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "swift/SwiftScan/DependencyScanningTool.h"
+#include "swift/DependencyScan/DependencyScanningTool.h"
 #include "swift/AST/DiagnosticEngine.h"
 #include "swift/AST/DiagnosticsFrontend.h"
 #include "swift/Basic/LLVMInitialize.h"
